requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
distlib==0.3.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-reversion==4.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.3
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
openpyxl==3.0.7
pbr==5.5.1
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stevedore==3.3.0
urllib3==1.26.6
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
